pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "term-dashboard"
version = "0.1.0"
description = "Terminal dashboard for real-time personal data sources."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "rich>=13.7.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "feedparser>=6.0.11",
  "google-api-python-client>=2.150.0",
  "google-auth>=2.34.0",
  "google-auth-oauthlib>=1.2.1",
]

[project.scripts]
termdash = "termdash.main:run"

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "respx>=0.21.1",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
